+++ src/prefect/tasks/azure/__init__.py
@@ -1,12 +1,14 @@
 """
 This module contains a collection of tasks for interacting with Azure resources.
-
-Note that all tasks require a Prefect Secret called `"AZ_CREDENTIALS"` that should be a JSON
-document with two keys: `"ACCOUNT_NAME"` and either `"ACCOUNT_KEY"` or `"SAS_TOKEN"`.
 """
 
 try:
     from prefect.tasks.azure.blobstorage import BlobStorageDownload, BlobStorageUpload
+    from prefect.tasks.azure.cosmosdb import (
+        CosmosDBCreateItem,
+        CosmosDBReadItems,
+        CosmosDBQueryItems,
+    )
 except ImportError:
     raise ImportError(
         'Using `prefect.tasks.azure` requires Prefect to be installed with the "azure" extra.'